[CCUBE-1058][FT] move element schema to its own file
